+++ src/js/init.js
@@ -1,6 +1,7 @@
 $(function () {
   $('input').typing();
   $('.as-input').asInput();
+  $('.input-date').inputDate();
   $('.chat').chat();
   $('[data-toggle="popover"]').popover();
   $('[data-toggle="tooltip"]').tooltip();