+++ src/js/types/date.js
@@ -1,0 +1,10 @@
+(function () {
+
+  Date.prototype.addDays = function (days) {
+    var date = new Date(this.valueOf());
+    date.setDate(date.getDate() + days);
+
+    return date;
+  }
+
+})();